Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() {
-int a,b,c;
+float a,b,c;
 float d;
-scanf("%d",&a);
-scanf("%d",&b);
-scanf("%d",&c);
+scanf("%f",&a);
+scanf("%f",&b);
+scanf("%f",&c);
 d = (a+b+c)/3;
 printf("Average: %.2f\n", d);
 return 0;
